+++ R/ggeffect.R
@@ -41,7 +41,7 @@
 #' plot(mydf)
 #'
 #' @importFrom purrr map
-#' @importFrom sjstats pred_vars resp_var
+#' @importFrom sjstats pred_vars resp_var model_family model_frame
 #' @importFrom dplyr if_else case_when bind_rows filter mutate
 #' @importFrom tibble as_tibble
 #' @importFrom sjmisc is_empty str_contains
@@ -70,7 +70,7 @@ ggeffect_helper <- function(model, terms, ci.lvl, x.as.factor, ...) {
   fitfram <- sjstats::model_frame(model)
 
   # get model family
-  faminfo <- get_glm_family(model)
+  faminfo <- sjstats::model_family(model)
 
   # create logical for family
   poisson_fam <- faminfo$is_pois